+++ src/Models/BotSetting.php
@@ -7,6 +7,12 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Stancl\Tenancy\Database\Concerns\BelongsToTenant;
 use TelegramBotEssentials\Essence\Models\Bot;
 
+/**
+ * @property int $id
+ * @property int $bot_id
+ * @property string $key
+ * @property string|null $value
+ */
 class BotSetting extends Model
 {
     use BelongsToTenant;